requirements.txt
# 웹 프레임워크
fastapi==0.109.0
uvicorn==0.27.0

# 파이썬 라이브러리 & 프레임워크
pyspark==3.5.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0

# 카프카
kafka-python==2.0.2 # Kafka를 사용하기 위한 클라이언트 라이브러리

# EK
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1

# DB
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# API Client
requests==2.31.0

# Async Support for FastAPI
aiohttp==3.9.1

# Jupyter Notebook
jupyter==1.0.0

# 이후 추가된 패키지
pyarrow==15.0.0  # PySpark 및 Pandas와 함께 사용하기 위해
python-dotenv==1.0.0  # 환경 변수 관리
pytest==7.4.4  # 테스트를 위해
black==23.12.1  # 코드 포맷팅
flake8==7.0.0  # 코드 린팅
